Delete config file only if it exists
git-svn-id: http://solar-empire.svn.sourceforge.net/svnroot/solar-empire@333 8f21f390-a30d-0410-b651-b82b18e590ef

+++ trunk/install/index.php
@@ -337,7 +337,9 @@ class swInstall
 	function configReset()
 	{
 		unset($_SESSION['configWritten']);
-		unlink(PATH_INC . '/config.inc.php');
+		if (file_exists(PATH_INC . '/config.inc.php')) {
+			unlink(PATH_INC . '/config.inc.php');
+		}
 	}
